New Changes
Created more pages to link to (just have to add links, in each page's nav bar). Styles on gaming secrets have been fixed as well as the theme button functionality.

New pages are still in need of style, so those will be next as well as getting theme change on each page.